conf/dj/requirements/dev.txt
-r base.txt
django-debug-toolbar
sqlparse